Kill components dir using inline dynamic imports
